feat: provide atomic cmsketch impl
